Always use legislation "effective end date" if present
This part of the query was copied from elsewhere,
but always using the earliest date doesn't seem to
make sense and leads to a lot of business rule
failures. Instead, we believe the correct behaviour
is to only use the original end date when the
effective end date is not present.